+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file. This change
 
 ## Unreleased
 
+### Fixed
+
+- use full name (namespace + name) for docker image tags.
+
 ## 0.10.187 (2023-08-20)
 
 ## 0.10.185 (2023-08-20)